deps/v8/test/inspector/debugger/es6-module-script-parsed-expected.txt

Debugger.scriptParsed and Debugger.scriptFailedToParse with ES6 module

Running test: testLoadedModulesOnDebuggerEnable
{
    method : Debugger.scriptParsed
    params : {
        embedderName : module1.js
        endColumn : 1
        endLine : 3
        executionContextId : <executionContextId>
        hasSourceURL : false
        hash : 2d91bcfe883a9f0a8da29373a3160d8ddc04ba4bc50f048a2cd1d80b9057ac03
        isLiveEdit : false
        isModule : true
        length : 39
        scriptId : <scriptId>
        scriptLanguage : JavaScript
        sourceMapURL :
        startColumn : 0
        startLine : 0
        url : module1.js
    }
}

Running test: testScriptEventsWhenDebuggerIsEnabled
{
    method : Debugger.scriptParsed
    params : {
        embedderName : module2.js
        endColumn : 1
        endLine : 3
        executionContextId : <executionContextId>
        hasSourceURL : false
        hash : 2d91bcfe883a9f0a8da29373a3160d8ddc04ba4bc50f048a2cd1d80b9057ac03
        isLiveEdit : false
        isModule : true
        length : 39
        scriptId : <scriptId>
        scriptLanguage : JavaScript
        sourceMapURL :
        startColumn : 0
        startLine : 0
        url : module2.js
    }
}
{
    method : Debugger.scriptFailedToParse
    params : {
        embedderName : module-with-syntax-error-2.js
        endColumn : 1
        endLine : 0
        executionContextId : <executionContextId>
        hasSourceURL : false
        hash : d10b36aa74a59bcf4a88185837f658afaf3646eff2bb16c3928d0e9335e945d2
        isModule : true
        length : 1
        scriptId : <scriptId>
        scriptLanguage : JavaScript
        sourceMapURL :
        startColumn : 0
        startLine : 0
        url : module-with-syntax-error-2.js
    }
}
